parse movie data from the JSON results

+++ app/src/main/java/com/example/henrytran/whattowatch/Movie.java
@@ -6,11 +6,11 @@ package com.example.henrytran.whattowatch;
 public class Movie {
     String title;
     String synopsis;
-    float rating;
+    double rating;
     String releaseDate;
     String posterPath;
 
-    public Movie (String title, String synopsis, float rating, String releaseDate, String posterPath) {
+    public Movie (String title, String synopsis, double rating, String releaseDate, String posterPath) {
         this.title = title;
         this.synopsis = synopsis;
         this.rating = rating;
